@@ -1,8 +1,7 @@
 name: qubic_wallet
 description: A Wallet for the Qubic Network created by Qubic-Hub.com
 publish_to: "none"
-version: 1.0.0+20
-
+version: 1.1.0+35
 environment:
   sdk: ">=3.0.0 <4.0.0"
 
@@ -11,22 +10,24 @@ dependencies:
   auto_size_text: ^3.0.0
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.5.0
+  flutter_localizations:
+    sdk: flutter
+  font_awesome_flutter: ^10.7.0
   from_css_color: ^2.0.0
-  google_fonts: ^5.1.0
-  intl:
-  json_path: ^0.6.2
+  google_fonts: ^6.2.1
+  intl: null
+  json_path: ^0.7.2
   page_transition: ^2.0.9
   shared_preferences: ^2.2.0
   timeago: ^3.5.0
   url_launcher: ^6.1.12
   mobx: ^2.2.0
-  flutter_mobx:
-  flutter_secure_storage: ^8.0.0
-  go_router: ^13.2.0
+  flutter_mobx: null
+  flutter_secure_storage: ^9.2.2
+  go_router: ^14.2.0
   get_it: ^7.6.0
-  form_builder_validators: ^9.0.0
-  flex_color_scheme:
+  form_builder_validators: ^11.0.0
+  flex_color_scheme: null
   dargon2_flutter: ^3.3.0
   flutter_form_builder: ^9.1.0
   persistent_bottom_nav_bar_v2: ^5.2.3
@@ -34,14 +35,14 @@ dependencies:
   flutter_inappwebview: ^6.0.0
   skeleton_text: ^3.0.1
   qr_flutter: ^4.1.0
-  share_plus: ^7.1.0
+  share_plus: ^9.0.0
   mobile_scanner: ^5.1.1
   flutter_multi_formatter: ^2.11.11
   pagination_flutter: ^0.0.8
   sticky_headers: "^0.3.0"
   equatable: ^2.0.5
   settings_ui: ^2.0.2
-  package_info_plus: ^4.2.0
+  package_info_plus: ^8.0.0
   local_auth: ^2.1.7
   http: ^1.2.1
   universal_platform: ^1.0.0+1
@@ -49,13 +50,15 @@ dependencies:
   window_manager: ^0.3.7
   dio: ^5.3.3
   percent_indicator: ^4.2.3
-  decimal: ^2.3.3
+  decimal: ^3.0.2
   animated_snack_bar: ^0.4.0
+  file_picker: ^8.0.6
+  downloadsfolder: ^1.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
   build_runner: ^2.4.4
   mobx_codegen: ^2.3.0
   flutter_launcher_icons: ^0.13.1
@@ -88,6 +91,7 @@ flutter_launcher_icons:
     image_path: "assets/launcher-icons/icon-1024x1024.png"
 
 flutter:
+  generate: true
   uses-material-design: true
   assets:
     - assets/images/
