@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:qubic_wallet/components/currency_label.dart';
-import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
-import 'package:qubic_wallet/styles/textStyles.dart';
-import 'package:skeleton_text/skeleton_text.dart';
+import 'package:qubic_wallet/styles/text_styles.dart';
 import 'package:intl/intl.dart';
 
 class AmountFormatted extends StatelessWidget {
