@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
-import 'package:qubic_wallet/globals.dart';
 import 'package:qubic_wallet/helpers/copy_to_clipboard.dart';
 import 'package:qubic_wallet/styles/themed_controls.dart';
 
@@ -14,7 +12,7 @@ class CopyButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return IconButton(
         onPressed: () async {
-          await copyToClipboard(copiedText);
+          await copyToClipboard(copiedText, context);
         },
         icon: LightThemeColors.shouldInvertIcon
             ? ThemedControls.invertedColors(
