@@ -4,7 +4,7 @@ import 'package:qubic_wallet/dtos/explorer_transaction_info_dto.dart';
 import 'package:qubic_wallet/helpers/transaction_UI_helpers.dart';
 
 import 'package:qubic_wallet/stores/application_store.dart';
-import 'package:qubic_wallet/styles/textStyles.dart';
+import 'package:qubic_wallet/styles/text_styles.dart';
 
 enum CardItem { delete, rename }
 
@@ -20,7 +20,7 @@ class ExplorerTransactionStatusItem extends StatelessWidget {
     return Row(mainAxisAlignment: MainAxisAlignment.start, children: [
       Icon(getTransactionStatusIcon(item.getStatus()),
           color: getTransactionStatusColor(item.getStatus()), size: 18),
-      Text(" ${getTransactionStatusText(item.getStatus())}",
+      Text(" ${getTransactionStatusText(item.getStatus(), context)}",
           style: TextStyles.labelTextSmall)
     ]);
   }
