@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
-import 'package:qubic_wallet/styles/textStyles.dart';
+import 'package:qubic_wallet/l10n/l10n.dart';
+import 'package:qubic_wallet/styles/text_styles.dart';
 import 'package:qubic_wallet/styles/themed_controls.dart';
 
 getAlertDialog(String title, String message,
@@ -13,7 +14,7 @@ getAlertDialog(String title, String message,
   if (secondaryButtonLabel != null) {
     assert(secondaryButtonFunction != null);
     actions.add(ConstrainedBox(
-        constraints: BoxConstraints(minWidth: 100),
+        constraints: const BoxConstraints(minWidth: 100),
         child: ThemedControls.transparentButtonBig(
           text: secondaryButtonLabel,
           onPressed: () async {
@@ -24,7 +25,7 @@ getAlertDialog(String title, String message,
         )));
   }
   actions.add(ConstrainedBox(
-      constraints: BoxConstraints(minWidth: 100),
+      constraints: const BoxConstraints(minWidth: 100),
       child: ThemedControls.primaryButtonBig(
         text: primaryButtonLabel,
         onPressed: () async {
@@ -42,15 +43,16 @@ getAlertDialog(String title, String message,
 }
 
 showAlertDialog(BuildContext context, String title, String message,
-    {Function? onOk}) {
-  // set up the button
-  Widget okButton = TextButton(
-    child: const Text("OK"),
+    {Function? primaryButtonFunction, String primaryButtonLabel = ""}) {
+  if (primaryButtonLabel.isEmpty) {
+    final l10n = l10nOf(context);
+    primaryButtonLabel = l10n.generalButtonOK;
+  }
+  Widget okButton = ThemedControls.primaryButtonBig(
+    text: primaryButtonLabel,
     onPressed: () async {
-      Navigator.of(context).pop();
-
-      if (onOk != null) {
-        onOk() {}
+      if (primaryButtonFunction != null) {
+        primaryButtonFunction();
       }
     },
   );
