@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:qubic_wallet/components/gradient_foreground.dart';
 import 'package:qubic_wallet/di.dart';
 import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
-import 'package:qubic_wallet/models/qubic_list_vm.dart';
 import 'package:qubic_wallet/stores/application_store.dart';
-import 'package:qubic_wallet/styles/edgeInsets.dart';
-import 'package:qubic_wallet/styles/textStyles.dart';
+import 'package:qubic_wallet/styles/edge_insets.dart';
+import 'package:qubic_wallet/styles/text_styles.dart';
 import 'package:qubic_wallet/styles/themed_controls.dart';
+import 'package:qubic_wallet/l10n/l10n.dart';
 
 class CreatePasswordSheet extends StatefulWidget {
   final Function() onAccept;
@@ -52,28 +51,29 @@ class _CreatePasswordSheetState extends State<CreatePasswordSheet> {
   }
 
   Widget getText() {
+    final l10n = l10nOf(context);
     return Row(children: [
       Expanded(
           child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
             ThemedControls.pageHeader(
-                headerText: "Please backup your password"),
-            Text(
-                "There is no way to retrieve your password if you forget it. Please make sure you have a backup of your password before proceeding.",
-                style: TextStyles.textLarge),
+                headerText: l10n.createPasswordSheetHeader),
+            Text(l10n.createPasswordSheetMessage, style: TextStyles.textLarge),
           ]))
     ]);
   }
 
 //transferNowHandler
   List<Widget> getButtons() {
+    final l10n = l10nOf(context);
+
     return [
       Expanded(
           child: ThemedControls.transparentButtonBigWithChild(
               child: Padding(
                   padding: const EdgeInsets.all(ThemePaddings.smallPadding),
-                  child: Text("Cancel",
+                  child: Text(l10n.generalButtonCancel,
                       textAlign: TextAlign.center,
                       style: TextStyles.transparentButtonText)),
               onPressed: widget.onReject)),
@@ -85,14 +85,14 @@ class _CreatePasswordSheetState extends State<CreatePasswordSheet> {
                   child: Padding(
                     padding:
                         const EdgeInsets.all(ThemePaddings.smallPadding + 3),
-                    child: Text("Proceed",
+                    child: Text(l10n.generalButtonProceed,
                         textAlign: TextAlign.center,
                         style: TextStyles.primaryButtonText),
                   ))
               : ThemedControls.primaryButtonBigDisabledWithChild(
                   child: Padding(
                   padding: const EdgeInsets.all(ThemePaddings.smallPadding + 3),
-                  child: Text("Proceed",
+                  child: Text(l10n.generalButtonProceed,
                       textAlign: TextAlign.center,
                       style: TextStyles.primaryButtonText),
                 ))),
@@ -104,10 +104,13 @@ class _CreatePasswordSheetState extends State<CreatePasswordSheet> {
       return;
     }
     widget.onAccept();
+    Navigator.pop(context);
   }
 
   @override
   Widget build(BuildContext context) {
+    final l10n = l10nOf(context);
+
     return Padding(
         padding: ThemeEdgeInsets.bottomSheetInsets,
         child: SingleChildScrollView(
@@ -121,7 +124,8 @@ class _CreatePasswordSheetState extends State<CreatePasswordSheet> {
                     ThemedControls.spacerVerticalNormal(),
                     const Divider(),
                     ListTile(
-                      contentPadding: EdgeInsets.symmetric(horizontal: 0.0),
+                      contentPadding:
+                          const EdgeInsets.symmetric(horizontal: 0.0),
                       leading: Checkbox(
                         value: hasAccepted,
                         onChanged: (value) {
@@ -130,7 +134,7 @@ class _CreatePasswordSheetState extends State<CreatePasswordSheet> {
                           });
                         },
                       ),
-                      title: const Text("Yes, I have backed up my password."),
+                      title: Text(l10n.createPasswordSheetCheckboxMessage),
                       onTap: () {
                         setState(() {
                           hasAccepted = !hasAccepted;
