@@ -1,56 +1,91 @@
+import 'dart:async';
+
 import 'package:animated_snack_bar/animated_snack_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:go_router/go_router.dart';
 import 'package:mobx/mobx.dart';
 import 'package:persistent_bottom_nav_bar_v2/persistent_bottom_nav_bar_v2.dart';
 import 'package:qubic_wallet/components/change_foreground.dart';
-import 'package:qubic_wallet/components/gradient_foreground.dart';
 import 'package:qubic_wallet/di.dart';
 import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
-import 'package:qubic_wallet/models/settings.dart';
-import 'package:qubic_wallet/pages/main/downloadCmdUtils.dart';
+import 'package:qubic_wallet/pages/main/download_cmd_utils.dart';
 import 'package:qubic_wallet/pages/main/tab_explorer.dart';
 import 'package:qubic_wallet/pages/main/tab_settings.dart';
 import 'package:qubic_wallet/pages/main/tab_transfers.dart';
 import 'package:qubic_wallet/pages/main/tab_wallet_contents.dart';
-import 'package:qubic_wallet/resources/qubic_cmd_utils.dart';
+import 'package:qubic_wallet/resources/qubic_cmd.dart';
 import 'package:qubic_wallet/stores/application_store.dart';
 import 'package:qubic_wallet/stores/qubic_hub_store.dart';
 import 'package:qubic_wallet/stores/settings_store.dart';
-import 'package:qubic_wallet/styles/textStyles.dart';
-import 'package:qubic_wallet/styles/themed_controls.dart';
+import 'package:qubic_wallet/styles/text_styles.dart';
 import 'package:qubic_wallet/timed_controller.dart';
 import 'package:universal_platform/universal_platform.dart';
-
-import '../../helpers/global_snack_bar.dart';
+import 'package:qubic_wallet/l10n/l10n.dart';
 
 class MainScreen extends StatefulWidget {
-  const MainScreen({super.key});
+  final int initialTabIndex;
+  const MainScreen({super.key, this.initialTabIndex = 0});
 
   @override
   // ignore: library_private_types_in_public_api
   _MainScreenState createState() => _MainScreenState();
 }
 
-class _MainScreenState extends State<MainScreen> {
+class _MainScreenState extends State<MainScreen> with WidgetsBindingObserver {
   late final PersistentTabController _controller;
   final _timedController = getIt<TimedController>();
   final QubicHubStore qubicHubStore = getIt<QubicHubStore>();
   final SettingsStore settingsStore = getIt<SettingsStore>();
   final ApplicationStore applicationStore = getIt<ApplicationStore>();
-
+  final QubicCmd qubicCmd = getIt<QubicCmd>();
   late final ReactionDisposer _disposeSnackbarAuto;
-  final _globalSnackBar = getIt<GlobalSnackBar>();
 
   late AnimatedSnackBar? errorBar;
   late AnimatedSnackBar? notificationBar;
 
+  Timer? _lockTimer;
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    if (state == AppLifecycleState.paused) {
+      // Lock the app immediately if the timeout is 0 (Immediately)
+      if (settingsStore.settings.autoLockTimeout == 0) {
+        applicationStore.signOut();
+      } else {
+        // Start the auto-lock timer when the app goes to background
+        _lockTimer = Timer(
+          Duration(minutes: settingsStore.settings.autoLockTimeout),
+          () {
+            // Lock the app
+            applicationStore.signOut();
+          },
+        );
+      }
+    } else if (state == AppLifecycleState.resumed) {
+      // Cancel the timer when the app is resumed
+      _lockTimer?.cancel();
+      if (!applicationStore.isSignedIn) {
+        context.go('/signIn');
+      }
+    }
+    if (state == AppLifecycleState.resumed) {
+      qubicCmd.reinitialize();
+    }
+  }
+
   @override
   void initState() {
     super.initState();
+    WidgetsBinding.instance.addObserver(this);
+
     _timedController.setupFetchTimer(true);
     _timedController.setupSlowTimer(true);
-    _controller = PersistentTabController(initialIndex: 0);
+    _controller = PersistentTabController(initialIndex: widget.initialTabIndex);
+    // _controller.jumpToTab(value);
+    _controller.addListener(() {
+      applicationStore.setCurrentTabIndex(_controller.index);
+    });
 
     if (!getIt.isRegistered<PersistentTabController>()) {
       getIt.registerSingleton<PersistentTabController>(_controller);
@@ -63,11 +98,14 @@ class _MainScreenState extends State<MainScreen> {
             ? applicationStore.globalError
             : applicationStore.globalError.substring(0, errorPos);
 
-        // AnimatedSnackBar.material(error,
-        //         type: AnimatedSnackBarType.error,
-        //         snackBarStrategy: StackSnackBarStrategy())
-        //     .show(context);
         if (error != "") {
+          //Error overriding for more than 15 accounts in wallet
+          if (error == "Failed to perform action. Server returned status 400") {
+            if (applicationStore.currentQubicIDs.length > 15) {
+              return;
+            }
+          }
+
           errorBar = AnimatedSnackBar(
               builder: ((context) {
                 return Ink(
@@ -90,6 +128,8 @@ class _MainScreenState extends State<MainScreen> {
               }),
               snackBarStrategy: RemoveSnackBarStrategy())
             ..show(context);
+
+          applicationStore.clearGlobalError();
         }
       }
 
@@ -130,21 +170,27 @@ class _MainScreenState extends State<MainScreen> {
   void dispose() {
     _timedController.stopFetchTimer();
     _disposeSnackbarAuto();
+
+    WidgetsBinding.instance.removeObserver(this);
+    _lockTimer?.cancel();
+
     super.dispose();
   }
 
   List<PersistentTabConfig> _tabs() {
+    final l10n = l10nOf(context);
+
     return [
       PersistentTabConfig(
           screen: Container(
               color: LightThemeColors.background,
-              child: const SafeArea(child: const TabWalletContents())),
+              child: const SafeArea(child: TabWalletContents())),
           item: ItemConfig(
             icon: ChangeForeground(
                 color: LightThemeColors.buttonBackground,
                 child: Image.asset("assets/images/tab-home.png")),
             inactiveIcon: Image.asset("assets/images/tab-home.png"),
-            title: ("Home"),
+            title: (l10n.appTabHome),
             textStyle: TextStyles.menuActive,
             activeForegroundColor: LightThemeColors.menuActive,
             inactiveForegroundColor: LightThemeColors.menuInactive,
@@ -158,7 +204,7 @@ class _MainScreenState extends State<MainScreen> {
                 color: LightThemeColors.buttonBackground,
                 child: Image.asset("assets/images/tab-transfers.png")),
             inactiveIcon: Image.asset("assets/images/tab-transfers.png"),
-            title: ("Transfers"),
+            title: (l10n.appTabTransfers),
             textStyle: TextStyles.menuActive,
             activeForegroundColor: LightThemeColors.menuActive,
             inactiveForegroundColor: LightThemeColors.menuInactive,
@@ -172,7 +218,7 @@ class _MainScreenState extends State<MainScreen> {
                 color: LightThemeColors.buttonBackground,
                 child: Image.asset("assets/images/tab-explorer.png")),
             inactiveIcon: Image.asset("assets/images/tab-explorer.png"),
-            title: ("Explorer"),
+            title: (l10n.appTabExplorer),
             textStyle: TextStyles.menuActive,
             activeForegroundColor: LightThemeColors.menuActive,
             inactiveForegroundColor: LightThemeColors.menuInactive,
@@ -186,7 +232,7 @@ class _MainScreenState extends State<MainScreen> {
                 color: LightThemeColors.buttonBackground,
                 child: Image.asset("assets/images/tab-settings.png")),
             inactiveIcon: Image.asset("assets/images/tab-settings.png"),
-            title: ("Settings"),
+            title: (l10n.appTabSettings),
             textStyle: TextStyles.menuActive,
             activeForegroundColor: LightThemeColors.menuActive,
             inactiveForegroundColor: LightThemeColors.menuInactive,
@@ -195,6 +241,12 @@ class _MainScreenState extends State<MainScreen> {
   }
 
   Widget getMain() {
+    if (applicationStore.hasStoredWalletSettings) {
+      _controller.jumpToTab(applicationStore.currentTabIndex);
+    } else {
+      _controller.jumpToTab(0);
+    }
+    // _controller.jumpToPreviousTab();
     return PersistentTabView(
       controller: _controller,
       navBarHeight: 60,
@@ -217,74 +269,6 @@ class _MainScreenState extends State<MainScreen> {
           true, // This needs to be true if you want to move up the screen when keyboard appears. Default is true.
       stateManagement: true, // Default is true.
     );
-    // return SafeArea(
-    //     child: Column(children: [
-    //   Observer(builder: (context) {
-    //     if (qubicHubStore.updateNeeded) {
-    //       return Container(
-    //           width: double.infinity,
-    //           color: Colors.red,
-    //           child: Column(children: [
-    //             TextButton(
-    //                 onPressed: () {},
-    //                 child: const Text(
-    //                     "This is an outdated version. Please update"))
-    //           ]));
-    //     }
-    //     return Container();
-    //   }),
-    //   Observer(builder: (context) {
-    //     if ((qubicHubStore.notice != null) && (qubicHubStore.notice != "")) {
-    //       return Container(
-    //           width: double.infinity,
-    //           color: Theme.of(context).cardColor,
-    //           child: Column(children: [
-    //             Flex(direction: Axis.horizontal, children: [
-    //               Expanded(
-    //                   child: Padding(
-    //                       padding:
-    //                           const EdgeInsets.all(ThemePaddings.smallPadding),
-    //                       child: Text(qubicHubStore.notice!,
-    //                           softWrap: true,
-    //                           maxLines: 3,
-    //                           style: Theme.of(context)
-    //                               .textTheme
-    //                               .labelSmall!
-    //                               .copyWith(
-    //                                   fontStyle: FontStyle.italic,
-    //                                   color: Theme.of(context)
-    //                                       .colorScheme
-    //                                       .secondary)))),
-    //               IconButton(
-    //                   onPressed: () {
-    //                     qubicHubStore.setNotice(null);
-    //                   },
-    //                   icon: Icon(Icons.close))
-    //             ])
-    //           ]));
-    //     }
-    //     return Container();
-    //   }),
-    //   Expanded(
-    //       child: PersistentTabView(
-    //     controller: _controller,
-    //     navBarBuilder: (navBarConfig) => Style1BottomNavBar(
-    //         navBarConfig: navBarConfig,
-    //         navBarDecoration: NavBarDecoration(
-    //             borderRadius: BorderRadius.circular(10.0),
-    //             color: LightThemeColors.navBg)),
-
-    //     tabs: _tabs(),
-
-    //     backgroundColor: LightThemeColors.navBg,
-
-    //     // Default is Colors.white.
-    //     handleAndroidBackButtonPress: true, // Default is true.
-    //     resizeToAvoidBottomInset:
-    //         true, // This needs to be true if you want to move up the screen when keyboard appears. Default is true.
-    //     stateManagement: true, // Default is true.
-    //   )),
-    //]));
   }
 
   @override
@@ -297,6 +281,7 @@ class _MainScreenState extends State<MainScreen> {
       if (UniversalPlatform.isDesktop && !settingsStore.cmdUtilsAvailable) {
         return const Scaffold(body: DownloadCmdUtils());
       }
+
       return getMain();
     });
   }
