@@ -6,28 +6,27 @@ import 'package:local_auth/local_auth.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:persistent_bottom_nav_bar_v2/persistent_bottom_nav_bar_v2.dart';
 import 'package:qubic_wallet/components/change_foreground.dart';
-import 'package:qubic_wallet/components/gradient_container.dart';
-import 'package:qubic_wallet/components/gradient_foreground.dart';
 import 'package:qubic_wallet/di.dart';
 import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
 import 'package:qubic_wallet/helpers/global_snack_bar.dart';
-import 'package:qubic_wallet/helpers/re_auth_dialog.dart';
 import 'package:qubic_wallet/pages/auth/erase_wallet_sheet.dart';
-import 'package:qubic_wallet/pages/main/wallet_contents/settings/about_wallet.dart';
 import 'package:qubic_wallet/pages/main/wallet_contents/settings/change_password.dart';
+import 'package:qubic_wallet/pages/main/wallet_contents/settings/export_wallet_vault.dart';
+import 'package:qubic_wallet/pages/main/wallet_contents/settings/join_community.dart';
 import 'package:qubic_wallet/pages/main/wallet_contents/settings/manage_biometics.dart';
+import 'package:qubic_wallet/pages/main/wallet_contents/settings/auto_lock_settings.dart';
 import 'package:qubic_wallet/resources/qubic_li.dart';
 import 'package:qubic_wallet/resources/secure_storage.dart';
 import 'package:qubic_wallet/stores/application_store.dart';
 import 'package:qubic_wallet/stores/explorer_store.dart';
 import 'package:qubic_wallet/stores/settings_store.dart';
 import 'package:qubic_wallet/stores/qubic_hub_store.dart';
-import 'package:qubic_wallet/styles/edgeInsets.dart';
-import 'package:qubic_wallet/styles/textStyles.dart';
-import 'package:qubic_wallet/styles/themed_controls.dart';
+import 'package:qubic_wallet/styles/edge_insets.dart';
+import 'package:qubic_wallet/styles/text_styles.dart';
 import 'package:qubic_wallet/timed_controller.dart';
 import 'package:settings_ui/settings_ui.dart';
 import 'package:universal_platform/universal_platform.dart';
+import 'package:qubic_wallet/l10n/l10n.dart';
 
 class TabSettings extends StatefulWidget {
   const TabSettings({super.key});
@@ -55,7 +54,7 @@ class _TabSettingsState extends State<TabSettings> {
   int itemsPerPage = 1000;
 
   BiometricType? biometricType; //The type of biometric available
-  String unlockLabel = "Unlock with Biometrics";
+  String settingsUnlockLabel = "";
   PackageInfo? packageInfo; // = await PackageInfo.fromPlatform();
   Widget icon = const Icon(Icons.fingerprint);
 
@@ -63,15 +62,24 @@ class _TabSettingsState extends State<TabSettings> {
   @override
   void initState() {
     super.initState();
+
     PackageInfo.fromPlatform().then((value) => setState(() {
           packageInfo = value;
         }));
+
     auth.canCheckBiometrics.then((value) {
+      final l10n = l10nOf(context);
+
+      // setting a default fallback value
+      settingsUnlockLabel =
+          l10n.settingsLabelManageBiometrics(l10n.generalBiometricTypeGeneric);
+
       auth.getAvailableBiometrics().then((value) {
         if ((value.contains(BiometricType.face)) && (biometricType == null)) {
           setState(() {
             biometricType = BiometricType.face;
-            unlockLabel = "Unlock with Face ID";
+            settingsUnlockLabel = l10n
+                .settingsLabelManageBiometrics(l10n.generalBiometricTypeFaceID);
             icon = Image.asset("assets/images/faceid.png");
           });
         }
@@ -79,14 +87,16 @@ class _TabSettingsState extends State<TabSettings> {
             (biometricType == null)) {
           setState(() {
             biometricType = BiometricType.fingerprint;
-            unlockLabel = "Unlock with Touch ID";
+            settingsUnlockLabel = l10n.settingsLabelManageBiometrics(
+                l10n.generalBiometricTypeTouchID);
             icon = const Icon(Icons.fingerprint);
           });
         }
         if ((value.contains(BiometricType.iris)) && (biometricType == null)) {
           setState(() {
             biometricType = BiometricType.iris;
-            unlockLabel = "Unlock with Iris";
+            settingsUnlockLabel = l10n
+                .settingsLabelManageBiometrics(l10n.generalBiometricTypeIris);
             icon = const Icon(Icons.remove_red_eye_outlined);
           });
         }
@@ -94,10 +104,12 @@ class _TabSettingsState extends State<TabSettings> {
           setState(() {
             biometricType = BiometricType.strong;
             if (UniversalPlatform.isWindows) {
-              unlockLabel = "Unlock with OS";
+              settingsUnlockLabel = l10n
+                  .settingsLabelManageBiometrics(l10n.generalBiometricTypeOS);
               icon = const Icon(Icons.security);
             } else {
-              unlockLabel = "Unlock with Biometrics";
+              settingsUnlockLabel = l10n.settingsLabelManageBiometrics(
+                  l10n.generalBiometricTypeGeneric);
               icon = const Icon(Icons.fingerprint);
             }
           });
@@ -112,15 +124,16 @@ class _TabSettingsState extends State<TabSettings> {
   }
 
   Widget getHeader() {
+    final l10n = l10nOf(context);
     return Padding(
         padding: const EdgeInsets.only(
             left: ThemePaddings.normalPadding,
             right: ThemePaddings.normalPadding,
             top: ThemePaddings.hugePadding,
             bottom: ThemePaddings.smallPadding),
-        child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: [Text("Settings", style: TextStyles.pageTitle)]));
+        child: Column(mainAxisAlignment: MainAxisAlignment.start, children: [
+          Text(l10n.appTabSettings, style: TextStyles.pageTitle)
+        ]));
   }
 
   Widget getBody() {
@@ -143,6 +156,8 @@ class _TabSettingsState extends State<TabSettings> {
   }
 
   Widget getSettings() {
+    final l10n = l10nOf(context);
+
     var theme = SettingsThemeData(
       settingsSectionBackground: LightThemeColors.cardBackground,
       //Theme.of(context).cardTheme.color,
@@ -168,17 +183,19 @@ class _TabSettingsState extends State<TabSettings> {
         lightTheme: theme,
         sections: [
           SettingsSection(
-            title: getSettingsHeader("Accounts and Data", true),
+            title: getSettingsHeader(l10n.settingsHeaderAccountsAndData, true),
             tiles: <SettingsTile>[
               SettingsTile.navigation(
-                leading: ChangeForeground(
-                    child: const Icon(Icons.lock),
-                    color: LightThemeColors.gradient1),
-                title: Text('Lock Wallet', style: TextStyles.textNormal),
+                leading: const ChangeForeground(
+                    color: LightThemeColors.gradient1, child: Icon(Icons.lock)),
+                title:
+                    Text(l10n.settingsLockWallet, style: TextStyles.textNormal),
                 trailing: Container(),
                 onPressed: (BuildContext context) {
                   appStore.reportGlobalError("");
                   appStore.reportGlobalNotification("");
+                  appStore.setCurrentTabIndex(
+                      0); // so after unlock, it goes to Home
                   appStore.signOut();
                   appStore.checkWalletIsInitialized();
                   timedController.stopFetchTimer();
@@ -186,10 +203,11 @@ class _TabSettingsState extends State<TabSettings> {
                 },
               ),
               SettingsTile.navigation(
-                leading: ChangeForeground(
-                    child: const Icon(Icons.cleaning_services_outlined),
-                    color: LightThemeColors.gradient1),
-                title: Text('Erase Wallet Data', style: TextStyles.textNormal),
+                leading: const ChangeForeground(
+                    color: LightThemeColors.gradient1,
+                    child: Icon(Icons.cleaning_services_outlined)),
+                title: Text(l10n.generalButtonEraseWalletData,
+                    style: TextStyles.textNormal),
                 trailing: Container(),
                 onPressed: (BuildContext context) async {
                   //MODAL TO CHECK IF USER AGREES
@@ -210,25 +228,41 @@ class _TabSettingsState extends State<TabSettings> {
                           timedController.stopFetchTimer();
                           Navigator.pop(context);
                           context.go("/signInNoAuth");
-                          _globalSnackBar
-                              .show("Wallet data erased from device");
+                          _globalSnackBar.show(
+                              l10n.generalSnackBarMessageWalletDataErased);
                         }, onReject: () async {
                           Navigator.pop(context);
                         }));
                       });
                 },
               ),
+              SettingsTile.navigation(
+                leading: const ChangeForeground(
+                    color: LightThemeColors.gradient1,
+                    child: Icon(Icons.file_present)),
+                title: Text(l10n.settingsLabelExportWalletVaultFile,
+                    style: TextStyles.textNormal),
+                trailing: Container(),
+                onPressed: (BuildContext context) async {
+                  pushScreen(context,
+                      screen: const ExportWalletVault(),
+                      withNavBar: false,
+                      pageTransitionAnimation:
+                          PageTransitionAnimation.cupertino);
+                },
+              ),
             ],
           ),
           SettingsSection(
-            title: getSettingsHeader("Security", true),
+            title: getSettingsHeader(l10n.settingsHeaderSecurity, true),
             tiles: <SettingsTile>[
               SettingsTile.navigation(
-                  leading: ChangeForeground(
-                      child: const Icon(Icons.password),
-                      color: LightThemeColors.gradient1),
+                  leading: const ChangeForeground(
+                      color: LightThemeColors.gradient1,
+                      child: Icon(Icons.password)),
                   trailing: getTrailingArrow(),
-                  title: Text('Change Password', style: TextStyles.textNormal),
+                  title: Text(l10n.settingsLabelChangePassword,
+                      style: TextStyles.textNormal),
                   onPressed: (BuildContext? context) async {
                     pushScreen(
                       context!,
@@ -239,13 +273,28 @@ class _TabSettingsState extends State<TabSettings> {
                     );
                   }),
               SettingsTile.navigation(
+                leading: const ChangeForeground(
+                    color: LightThemeColors.gradient1,
+                    child: Icon(Icons.lock_clock)),
+                title: Text(l10n.settingsLabelAutlock,
+                    style: TextStyles.textNormal),
+                trailing: getTrailingArrow(),
+                onPressed: (BuildContext? context) async {
+                  pushScreen(context!,
+                      screen: AutoLockSettings(),
+                      withNavBar: false,
+                      pageTransitionAnimation:
+                          PageTransitionAnimation.cupertino);
+                },
+              ),
+              SettingsTile.navigation(
                 leading: ChangeForeground(
                     color: LightThemeColors.gradient1, child: icon),
                 trailing: getTrailingArrow(),
-                title: Text(unlockLabel, style: TextStyles.textNormal),
+                title: Text(settingsUnlockLabel, style: TextStyles.textNormal),
                 value: Observer(builder: (context) {
                   return settingsStore.settings.biometricEnabled
-                      ? Text("Enabled",
+                      ? Text(l10n.generalLabelEnabled,
                           style: Theme.of(context)
                               .textTheme
                               .bodySmall!
@@ -253,7 +302,7 @@ class _TabSettingsState extends State<TabSettings> {
                                   color:
                                       Theme.of(context).colorScheme.secondary,
                                   fontFamily: ThemeFonts.secondary))
-                      : const Text("Disabled");
+                      : Text(l10n.generalLabelDisabled);
                 }),
                 onPressed: (BuildContext? context) {
                   pushScreen(
@@ -267,12 +316,29 @@ class _TabSettingsState extends State<TabSettings> {
             ],
           ),
           SettingsSection(
-              title: getSettingsHeader("Other", true),
+              title: getSettingsHeader(l10n.settingsHeaderAbout, true),
               tiles: <SettingsTile>[
                 SettingsTile.navigation(
-                  leading: ChangeForeground(
-                      child: const Icon(Icons.account_balance_wallet_outlined),
-                      color: LightThemeColors.gradient1),
+                  leading: const ChangeForeground(
+                      color: LightThemeColors.gradient1,
+                      child: Icon(Icons.people)),
+                  title: Text(l10n.settingsLabelJoinCommunity,
+                      style: TextStyles.textNormal),
+                  trailing: Container(),
+                  onPressed: (BuildContext? context) async {
+                    pushScreen(
+                      context!,
+                      screen: const JoinCommunity(),
+                      withNavBar: false, // OPTIONAL VALUE. True by default.
+                      pageTransitionAnimation:
+                          PageTransitionAnimation.cupertino,
+                    );
+                  },
+                ),
+                SettingsTile.navigation(
+                  leading: const ChangeForeground(
+                      color: LightThemeColors.gradient1,
+                      child: Icon(Icons.account_balance_wallet_outlined)),
                   trailing: Observer(builder: (BuildContext context) {
                     if (qubicHubStore.updateAvailable) {
                       return const Icon(Icons.info, color: Colors.red);
@@ -283,10 +349,11 @@ class _TabSettingsState extends State<TabSettings> {
                       mainAxisAlignment: MainAxisAlignment.start,
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        Text('Wallet Info', style: TextStyles.textNormal),
+                        Text(l10n.settingsLabelWalletInfo,
+                            style: TextStyles.textNormal),
                         Observer(builder: (BuildContext context) {
                           if (qubicHubStore.versionInfo == null) {
-                            return const Text("Loading...");
+                            return Text(l10n.generalLabelLoading);
                           }
                           return Row(
                               mainAxisAlignment: MainAxisAlignment.spaceBetween,
