@@ -5,9 +5,10 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
-import 'package:crypto/crypto.dart' as crypto;
-import 'package:flutter/services.dart' show rootBundle;
-import 'package:qubic_wallet/config.dart';
+import 'package:flutter/services.dart' show Uint8List;
+import 'package:qubic_wallet/globals/localization_manager.dart';
+import 'package:qubic_wallet/models/qubic_import_vault_seed.dart';
+import 'package:qubic_wallet/models/qubic_vault_export_seed.dart';
 
 /// A class that handles the secure storage of the wallet. The wallet is stored in the secure storage of the device
 /// The wallet password is encrypted using Argon2
@@ -23,26 +24,32 @@ class QubicJs {
       return;
     }
     InAppWebView = HeadlessInAppWebView(
-        onWebViewCreated: (WVcontroller) async {
-          WVcontroller.loadFile(
-              assetFilePath: "assets/qubic_js/qubic-helper-html-3_0_1.html");
+      onWebViewCreated: (WVcontroller) async {
+        WVcontroller.loadFile(
+            assetFilePath: "assets/qubic_js/qubic-helper-html-3_0_5.html");
 
-          controller = WVcontroller;
-        },
-        onConsoleMessage: (controller, consoleMessage) {
-          debugPrint(consoleMessage.toString());
-        },
-        onReceivedError: (controller, request, error) =>
-            {debugPrint(error.toString())},
-        onLoadStart: (controller, url) {},
-        onLoadStop: (controller, url) async {
-          isReady = true;
-        });
+        controller = WVcontroller;
+      },
+      onConsoleMessage: (controller, consoleMessage) {
+        debugPrint(consoleMessage.toString());
+      },
+      onReceivedError: (controller, request, error) =>
+          {debugPrint(error.toString())},
+      onLoadStart: (controller, url) {},
+      onLoadStop: (controller, url) async {
+        isReady = true;
+      },
+    );
 
     await InAppWebView!.run();
     while (controller == null) {
       sleep(const Duration(milliseconds: 100));
     }
+  }
+
+  reInitialize() async {
+    disposeController();
+    await initialize();
   }
 
   bool isReady = false;
@@ -87,11 +94,12 @@ class QubicJs {
         await controller!.callAsyncJavaScript(functionBody: functionBody);
 
     if (result == null) {
-      throw Exception("Error trying to create asset transfer transcation");
+      throw Exception(LocalizationManager.instance.appLocalization
+          .cmdErrorCreatingAssetTransferTransactionGeneric);
     }
     if (result.error != null) {
-      throw Exception(
-          "Error trying to create asset transfer transcation: ${result.error}");
+      throw Exception(LocalizationManager.instance.appLocalization
+          .cmdErrorCreatingAssetTransferTransaction(result.error ?? ""));
     }
     return result.value['transaction'];
   }
@@ -107,10 +115,12 @@ class QubicJs {
         await controller!.callAsyncJavaScript(functionBody: functionBody);
 
     if (result == null) {
-      throw Exception("Error trying to create transcation");
+      throw Exception(LocalizationManager
+          .instance.appLocalization.cmdErrorCreatingTransferTransactionGeneric);
     }
     if (result.error != null) {
-      throw Exception("Error trying to create transcation: ${result.error}");
+      throw Exception(LocalizationManager.instance.appLocalization
+          .cmdErrorCreatingTransferTransaction(result.error ?? ""));
     }
     return result.value['transaction'];
   }
@@ -121,11 +131,112 @@ class QubicJs {
             "return await runBrowser('createPublicId','${seed.replaceAll("'", "\\'")}');");
 
     if (result == null) {
-      throw Exception('Error getting public id from seed: Generic error');
+      throw Exception(LocalizationManager
+          .instance.appLocalization.cmdErrorGettingPublicIdFromSeedGeneric);
     }
     if (result.error != null) {
-      throw Exception('Error getting public id from seed:  ${result.error!}');
+      throw Exception(LocalizationManager.instance.appLocalization
+          .cmdErrorGettingPublicIdFromSeed(result.error ?? ""));
     }
     return result.value['publicId'];
+  }
+
+  /// Return base64  vault file
+  Future<Uint8List> createVaultFile(
+      String password, List<QubicVaultExportSeed> seeds) async {
+    CallAsyncJavaScriptResult? result = await controller!.callAsyncJavaScript(
+        functionBody:
+            "return await runBrowser('wallet.createVaultFile','${password.replaceAll("'", "\\'")}','${jsonEncode(seeds.map((e) => e.toJson()).toList()).replaceAll("'", "\\'")}')");
+    if (result == null) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.cmdErrorCreatingVaultFileGeneric);
+    }
+    if (result.value['status'] == 'error') {
+      throw Exception(LocalizationManager.instance.appLocalization
+          .exportWalletVaultErrorGeneralMessage(result.value['error'] ?? ""));
+    }
+    if (result.error != null) {
+      throw Exception(LocalizationManager.instance.appLocalization
+          .exportWalletVaultErrorGeneralMessage(result.error ?? ""));
+    }
+    if (result.value['base64'] == null) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.cmdErrorCreatingVaultFileGeneratedIsEmpty);
+    }
+    return base64Decode(result.value['base64']!);
+  }
+
+  Future<List<QubicImportVaultSeed>> importVault(
+      String password, String baseFileContents) async {
+    List<QubicImportVaultSeed>? seeds;
+    List<dynamic>? parsedSeeds;
+    CallAsyncJavaScriptResult? result = await controller!.callAsyncJavaScript(
+        functionBody:
+            "return await runBrowser('wallet.importVault','${password.replaceAll("'", "\\'")}','${baseFileContents.replaceAll("'", "\\'")}')");
+
+    if (result == null) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.importVaultFileGenericError);
+    }
+    if (result.value['status'] == 'error') {
+      if (result.value['error'] == "Could not parse seeds JSON") {
+        throw Exception(LocalizationManager
+            .instance.appLocalization.importVaultFileOrPasswordError);
+      }
+      throw Exception(LocalizationManager.instance.appLocalization
+          .importVaultFileErrorGeneralMessage(result.value['error'] ?? ""));
+    }
+    if (result.error != null) {
+      throw Exception(LocalizationManager.instance.appLocalization
+          .importVaultFileErrorGeneralMessage(result.error ?? ""));
+    }
+    if (result.value['seeds'] == null) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.importVaultFileSeedsNullError);
+    }
+    if (result.value['seeds'].toString().isEmpty) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.importVaultErrorNoAccountsFound);
+    }
+
+    try {
+      parsedSeeds = result.value['seeds'];
+    } catch (e) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.importVaultFileSeedsMalformedError);
+    }
+
+    if (parsedSeeds == null) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.importVaultFileParsedSeedsNullError);
+    }
+
+    if (parsedSeeds.isEmpty) {
+      throw Exception(LocalizationManager
+          .instance.appLocalization.importVaultErrorNoAccountsFound);
+    }
+
+    seeds = <QubicImportVaultSeed>[];
+    var i = 0;
+    for (var seed in parsedSeeds) {
+      if (seed['alias'] == null) {
+        throw Exception(LocalizationManager.instance.appLocalization
+            .importVaultFileAccountMissingName(i));
+      }
+      if (seed['publicId'] == null) {
+        throw Exception(LocalizationManager.instance.appLocalization
+            .importVaultFileAccountMissingPublicId(i));
+      }
+      if (seed['seed'] == null) {
+        throw Exception(LocalizationManager.instance.appLocalization
+            .importVaultFileAccountMissingSeed(i));
+      }
+
+      seeds.add(QubicImportVaultSeed(
+          seed['alias'], seed['publicId'], seed['seed'] ?? ""));
+      i++;
+    }
+
+    return seeds;
   }
 }
